Split Vuex Store in Multiple Files
