+++ src/store/index.js
@@ -1,6 +1,6 @@
 import Vuex from 'vuex'
 import Vue from 'vue'
-import shop from '@/api/shop'
+import actions from './actions'
 
 Vue.use(Vuex)
 
@@ -38,45 +38,7 @@ export default new Vuex.Store({
         }
     },
 
-    actions: {
-        fetchProducts ({commit}) {
-
-            return new Promise((resolve, reject) => {
-                // make the call
-                shop.getProducts( products => {
-                    commit('setProducts', products)
-                    resolve()
-                })
-
-            });
-
-        },
-
-        addProductToCart ({state, getters, commit}, product) {
-            if (getters.productIsInStock(product)) {
-                const cartItem = state.cart.find( item => item.id === product.id )
-                if (!cartItem) {
-                    commit('pushProductToCart', product.id)
-                } else {
-                    commit('incrementItemQuantity', cartItem)
-                }
-                commit('decrementProductInventory', product)
-            }
-        },
-
-        checkout ({state, commit}) {
-            shop.buyProducts(
-                state.cart,
-                () => {
-                    commit('emptyCart')
-                    commit('setCheckoutStatus', 'success')
-                },
-                () => {
-                    commit('setCheckoutStatus', 'fail')
-                }
-            )
-        }
-    },
+    actions,
 
     mutations: {
         setProducts (state, products) {
